+++ NEWS.md
@@ -1,4 +1,16 @@
+# zoltr 0.2.2
+- Added automatic re-authentication of expired tokens
+
+
+- Made changes to avoid CRAN NOTE:
+# zoltr 0.2.1
+- Made changes to avoid CRAN NOTE:
+
+    checking dependencies in R code ... NOTE
+    Namespaces in Imports field not imported from:
+    ‘jsonlite’ ‘mockery’ ‘readr’ ‘webmockr’
+    All declared Imports should be used.
+
+
 # zoltr 0.2.0
-
-- Initial private beta release!
-
+- Initial private beta release